Modularize service worker cache
